Updates Localytics to v5
The one change here is that when configuring the appKey, the Localytics integrate method
takes an additional argument to add flexibility on when/how to upload user data.

Localytics will default to uploading data periodically based on the state of a user's network connection.
The default behavior occurs when passing `nil` to `withLocalyticsOptions`.

I think we can ship this update and consider supporting the ability to integrate with Localytics options
in another release.